start removing casts from MIN and MAX

+++ src/sixel.c
@@ -597,8 +597,8 @@ sixel_parser_parse(SixelState *sixel_state, uchar *p, int32 len) {
                      * parsing the last sixel line */
                     sy = (sy + 5) / 6*6;
 
-                    sx = (int32)MIN(sx, DECSIXEL_WIDTH_MAX);
-                    sy = (int32)MIN(sy, DECSIXEL_HEIGHT_MAX);
+                    sx = MIN(sx, DECSIXEL_WIDTH_MAX);
+                    sy = MIN(sy, DECSIXEL_HEIGHT_MAX);
 
                     if (sixel_image_buffer_resize(sixel_image, sx, sy) < 0) {
                         error("Error in sixel_image_buffer_resize.\n");
